Move DoNotMock to internal and make it public for internal reuse
